split websocket into its own file
