+++ engine/server/engine/server/websocket_api_handler.go
@@ -1,13 +1,10 @@
 package server
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"time"
 
-	"github.com/gorilla/websocket"
 	user_service "github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis/engine/server/engine/centralized_logs"
 	"github.com/kurtosis-tech/kurtosis/engine/server/engine/centralized_logs/client_implementations/persistent_volume/log_file_manager"
@@ -23,19 +20,6 @@ import (
 
 	rpc_api "github.com/kurtosis-tech/kurtosis/api/golang/core/kurtosis_core_rpc_api_bindings"
 	api_type "github.com/kurtosis-tech/kurtosis/api/golang/http_rest/api_types"
-)
-
-const (
-	wsReadBufferSize  = 1024
-	wsWriteBufferSize = 1024
-	maxMessageSize    = 512
-	// Time allowed to write a message to the peer.
-	writeWait = 10 * time.Second
-	// Time allowed to read the next pong message from the peer.
-	pongWait = 60 * time.Second
-	// Send pings to peer with this period. Must be less than pongWait.
-	// nolint:gomnd
-	pingPeriod = (pongWait * 9) / 10
 )
 
 type WebSocketRuntime struct {
@@ -179,7 +163,7 @@ func streamStarlarkLogsWithWebsocket[T any](ctx echo.Context, cors cors.Cors, st
 		return
 	}
 
-	wsPump, err := NewWebsocketPump[api_type.StarlarkRunResponseLine](ctx, cors)
+	wsPump, err := streaming.NewWebsocketPump[api_type.StarlarkRunResponseLine](ctx, cors)
 	if err != nil {
 		logrus.WithError(err).Error("Failed to start websocket connection")
 		ctx.Response().Header().Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
@@ -196,19 +180,19 @@ func streamStarlarkLogsWithWebsocket[T any](ctx echo.Context, cors cors.Cors, st
 		return
 	}
 	defer wsPump.Close()
-	go wsPump.startPumping()
+	go wsPump.StartPumping()
 
 	found, err := streamerPool.Consume(streaming.StreamerUUID(streamerUUID), func(logline *rpc_api.StarlarkRunResponseLine) error {
 		response, err := to_http.ToHttpStarlarkRunResponseLine(logline)
 		if err != nil {
 			return stacktrace.Propagate(err, "Failed to convert value of type `%T` to http", logline)
 		}
-		wsPump.pumpMessage(response)
+		wsPump.PumpMessage(response)
 		return nil
 	})
 
 	if !found {
-		wsPump.pumpResponseInfo(&notFoundErr)
+		wsPump.PumpResponseInfo(&notFoundErr)
 	}
 
 	if err != nil {
@@ -221,7 +205,7 @@ func streamStarlarkLogsWithWebsocket[T any](ctx echo.Context, cors cors.Cors, st
 			Message: fmt.Sprintf("Log streaming '%s' failed while sending the data", streamerUUID),
 			Code:    http.StatusInternalServerError,
 		}
-		wsPump.pumpResponseInfo(&streamingErr)
+		wsPump.PumpResponseInfo(&streamingErr)
 	}
 }
 
@@ -275,7 +259,7 @@ func streamStarlarkLogsWithHTTP[T any](ctx echo.Context, streamerPool streaming.
 }
 
 func streamServiceLogsWithWebsocket(ctx echo.Context, cors cors.Cors, streamer streaming.ServiceLogStreamer) {
-	wsPump, err := NewWebsocketPump[api_type.ServiceLogs](ctx, cors)
+	wsPump, err := streaming.NewWebsocketPump[api_type.ServiceLogs](ctx, cors)
 	if err != nil {
 		logrus.WithError(err).Error("Failed to start websocket connection")
 		writeResponseInfo(ctx, api_type.ResponseInfo{
@@ -286,10 +270,10 @@ func streamServiceLogsWithWebsocket(ctx echo.Context, cors cors.Cors, streamer s
 		return
 	}
 	defer wsPump.Close()
-	go wsPump.startPumping()
+	go wsPump.StartPumping()
 
 	err = streamer.Consume(func(logline *api_type.ServiceLogs) error {
-		wsPump.pumpMessage(logline)
+		wsPump.PumpMessage(logline)
 		return nil
 	})
 
@@ -303,7 +287,7 @@ func streamServiceLogsWithWebsocket(ctx echo.Context, cors cors.Cors, streamer s
 			Message: "Log streaming failed while sending the data",
 			Code:    http.StatusInternalServerError,
 		}
-		wsPump.pumpResponseInfo(&streamingErr)
+		wsPump.PumpResponseInfo(&streamingErr)
 	}
 }
 
@@ -333,96 +317,4 @@ func streamServiceLogsWithHTTP(ctx echo.Context, streamer streaming.ServiceLogSt
 			logrus.WithError(err).Errorf("Failed to send value via websocket")
 		}
 	}
-}
-
-type WebsocketPump[T interface{}] struct {
-	websocket  *websocket.Conn
-	inputChan  chan *T
-	infoChan   chan *api_type.ResponseInfo
-	ctx        context.Context
-	cancelFunc context.CancelFunc
-}
-
-func NewWebsocketPump[T interface{}](ctx echo.Context, cors cors.Cors) (*WebsocketPump[T], error) {
-	// nolint: exhaustruct
-	upgrader := websocket.Upgrader{
-		ReadBufferSize:  wsReadBufferSize,
-		WriteBufferSize: wsWriteBufferSize,
-		CheckOrigin:     cors.OriginAllowed,
-	}
-
-	conn, err := upgrader.Upgrade(ctx.Response(), ctx.Request(), nil)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "Failed to upgrade http connection to websocket")
-	}
-
-	ctxWithCancel, cancelFunc := context.WithCancel(context.Background())
-
-	return &WebsocketPump[T]{
-		websocket:  conn,
-		inputChan:  make(chan *T),
-		ctx:        ctxWithCancel,
-		cancelFunc: cancelFunc,
-	}, nil
-}
-
-func (pump WebsocketPump[T]) startPumping() {
-	ticker := time.NewTicker(pingPeriod)
-	defer func() {
-		ticker.Stop()
-		pump.websocket.Close()
-	}()
-
-	logrus.WithFields(logrus.Fields{
-		"pongWait":       pongWait,
-		"pingPeriod":     pingPeriod,
-		"maxMessageSize": maxMessageSize,
-	}).Debug("Started keep alive process for websocket connection.")
-
-	pump.websocket.SetReadLimit(maxMessageSize)
-	if err := pump.websocket.SetReadDeadline(time.Now().Add(pongWait)); err != nil {
-		logrus.WithError(err).Error("Failed to set Pong wait time")
-		return
-	}
-	// nolint:errcheck
-	pump.websocket.SetPongHandler(func(string) error { return pump.websocket.SetReadDeadline(time.Now().Add(pongWait)) })
-
-	for {
-		select {
-		case <-ticker.C:
-			if err := pump.websocket.SetWriteDeadline(time.Now().Add(writeWait)); err != nil {
-				logrus.Debug("Websocket connection is likely closed, exiting keep alive process")
-				return
-			}
-			if err := pump.websocket.WriteMessage(websocket.PingMessage, nil); err != nil {
-				logrus.Debug("Websocket connection is likely closed, exiting keep alive process")
-				return
-			}
-		case msg := <-pump.inputChan:
-			if err := pump.websocket.WriteJSON(msg); err != nil {
-				logrus.WithError(stacktrace.Propagate(err, "Failed to send value of type `%T` via websocket", msg)).Errorf("Failed to write message to websocket, closing it.")
-				return
-			}
-		case msg := <-pump.infoChan:
-			if err := pump.websocket.WriteJSON(msg); err != nil {
-				logrus.WithError(stacktrace.Propagate(err, "Failed to send value of type `%T` via websocket", msg)).Errorf("Failed to write message to websocket, closing it.")
-				return
-			}
-		case <-pump.ctx.Done():
-			logrus.Debug("Websocket pumper has been asked to close, closing it.")
-			return
-		}
-	}
-}
-
-func (pump *WebsocketPump[T]) pumpResponseInfo(msg *api_type.ResponseInfo) {
-	pump.infoChan <- msg
-}
-
-func (pump *WebsocketPump[T]) pumpMessage(msg *T) {
-	pump.inputChan <- msg
-}
-
-func (pump *WebsocketPump[T]) Close() {
-	pump.cancelFunc()
 }
